requirements.txt
pyimgui
pygame < 2.0.0